refactor(rooms): remove ngrx-effects for Rooms
