+++ src/app/app.state.ts
@@ -1,10 +1,9 @@
 import { Room } from './interfaces/room.model';
-import { Rooms } from './interfaces/rooms.model';
 import { Player } from './interfaces/player.model';
 import { WordsState } from './interfaces/words.interfaces';
 
 export interface AppState {
-    rooms: Rooms;
+    rooms: Room[];
     room: Room;
     player: Player;
     words: WordsState;
