--- src/app/interfaces/rooms.model.ts
@@ -1,6 +1,0 @@
-import { Room } from './room.model';
-
-export interface Rooms {
-    items: Room[];
-    loading: boolean;
-}
