--- src/app/rooms/rooms.effects.ts
@@ -1,40 +1,0 @@
-import { Injectable } from '@angular/core';
-import { Actions, Effect } from '@ngrx/effects';
-import { Action } from '@ngrx/store';
-import { AngularFireDatabase, FirebaseListObservable } from 'angularfire2/database';
-
-import { GET_ROOMS, GetRoomsSuccess } from './rooms.actions';
-
-import { Observable } from 'rxjs';
-
-import { map, mergeMap } from 'rxjs/operators';
-
-import capitalize from 'lodash-es/capitalize';
-
-const WORDS = [];
-
-@Injectable()
-export class RoomsEffects {
-
-    constructor(private actions: Actions,
-                private db: AngularFireDatabase) {
-    }
-
-    @Effect()
-    getRooms: Observable<Action> = this.actions.ofType(GET_ROOMS)
-        .pipe(
-            mergeMap(() => this._getRooms()),
-            map((rooms: any) => GetRoomsSuccess(rooms)),
-        );
-
-    private _getRooms(): FirebaseListObservable<any> {
-        return this.db.list('/rooms');
-    }
-
-    public submitMonikerWords() {
-        WORDS.map((word: string) => {
-            const wordToPush = word.split(' ').map((o: string) => capitalize(o)).join(' ');
-            this.db.list('/words/monikers').push(wordToPush);
-        });
-    }
-}
